fix android takePicture

+++ src/android/CameraPreview.java
@@ -222,8 +222,8 @@ public class CameraPreview extends CordovaPlugin implements CameraActivity.Camer
 
     try {
       DisplayMetrics metrics = cordova.getActivity().getResources().getDisplayMetrics();
-      double width = (double) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, args.getInt(0), metrics);
-      double height = (double) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, args.getInt(1), metrics);
+      int width = (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, args.getInt(0), metrics);
+      int height = (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, args.getInt(1), metrics);
       int quality = args.getInt(2);
       fragment.takePicture(width, height, quality);
